requirements.txt
numpy
langchain
langchain-community
faiss-cpu
openai
transformers
sentence-transformers
streamlit